demo/chat/requirements.txt
openai
fastapi
uvicorn
httpx
python-multipart
sqlite3
hashlib
chardet
fpdf2
python-docx
reportlab
python-pptx
pandas
matplotlib
seaborn
scikit-learn
numpy
openpyxl
pdfplumber
pdfminer.six
pypdfium2
sqlalchemy
pymysql
psycopg2-binary
pymssql
